Rename ChatMessage to ChatPrompt
